Order fixed

+++ TMD.Models/DomainModels/OrderItem.cs
@@ -10,7 +10,7 @@
         public decimal AmountGiven { get; set; }
         public decimal SalePrice { get; set; }
         public decimal MinSalePriceAllowed { get; set; }
-        public byte Discount { get; set; }
+        public int Discount { get; set; }
         public string Comments { get; set; }
         public System.DateTime RecCreatedDate { get; set; }
         public System.DateTime RecLastUpdatedDate { get; set; }
